requirements.txt
absl-py==0.7.0
astor==0.7.1
atari-py==0.1.7
certifi==2018.11.29
chardet==3.0.4
Click==7.0
cloudpickle==0.8.0
cycler==0.10.0
dill==0.2.9
future==0.17.1
gast==0.2.2
glob2==0.6
grpcio==1.19.0
gym==0.12.0
h5py==2.9.0
idna==2.8
joblib==0.13.2
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
Markdown==3.0.1
matplotlib==3.0.3
mpi4py==3.0.1
numpy==1.16.2
opencv-python==4.0.0.21
pandas==0.24.1
Pillow==5.4.1
progressbar2==3.39.2
protobuf==3.7.0
pyglet==1.3.2
PyOpenGL==3.1.0
pyparsing==2.3.1
python-dateutil==2.8.0
python-utils==2.3.0
pytz==2018.9
PyYAML==3.13
pyzmq==18.0.0
requests==2.21.0
scipy==1.2.1
seaborn==0.9.0
six==1.12.0
tensorboard==1.12.2
tensorflow-gpu==1.12.0
termcolor==1.1.0
tqdm==4.31.1
urllib3==1.24.1
Werkzeug==0.14.1
zmq==0.0.0
